+++ sample/sample3-nested-objects/Tag.ts
@@ -1,8 +1,0 @@
-import { Contains, IsInt, Length, IsEmail, IsFQDN, IsDate } from '../../src/decorator/decorators';
-
-export class Tag {
-  @Length(10, 20, {
-    message: 'Tag is too short or long',
-  })
-  name: string;
-}