+++ sample/sample5-schemas/Post.ts
@@ -1,9 +1,0 @@
-export class Post {
-  title: string;
-  text: string;
-  rating: number;
-  email: string;
-  site: string;
-  createDate: Date;
-  tags: string[];
-}