+++ sample/sample7-inheritance-support/BaseContent.ts
@@ -1,6 +1,0 @@
-import { IsEmail } from '../../src/decorator/decorators';
-
-export class BaseContent {
-  @IsEmail()
-  email: string;
-}